refactor: 调整全部拼写错误 lable => label
