+++ src/components/LayerAttribute/BubbleLayerStyle/schema.ts
@@ -1,5 +1,5 @@
 import fillColorCollapse from '../common-schema/fill-color-collapse';
-import lableCollapse from '../common-schema/lable-collapse';
+import labelCollapse from '../common-schema/label-collapse';
 import otherAttributesCollapse from '../common-schema/other-attributes-collapse';
 import pointRadius from '../common-schema/point-radius-collapse';
 import strokeCollapse from '../common-schema/stroke-collapse';
@@ -12,7 +12,7 @@ export default (fieldList: FieldSelectOptionType[] = [], ribbonList: string[][] 
       collapseItem_fillColor: fillColorCollapse(fieldList, ribbonList),
       collapseItem_stroke: strokeCollapse(),
       collapseItem_fillradius: pointRadius(fieldList),
-      collapseItem_labelstyle: lableCollapse(fieldList),
+      collapseItem_labelstyle: labelCollapse(fieldList),
       // collapseItem_animate: animateCollapse(),
       collapseItem_other: otherAttributesCollapse(),
     },
