+++ src/components/LayerAttribute/ChoroplethLayerStyle/schema.ts
@@ -1,5 +1,5 @@
 import fillColorCollapse from '../common-schema/fill-color-collapse';
-import lableCollapse from '../common-schema/lable-collapse';
+import labelCollapse from '../common-schema/label-collapse';
 import otherAttributesCollapse from '../common-schema/other-attributes-collapse';
 import strokeCollapse from '../common-schema/stroke-collapse';
 import type { FieldSelectOptionType } from '../types';
@@ -10,7 +10,7 @@ export default (fieldList: FieldSelectOptionType[] = [], ribbonList: string[][] 
     properties: {
       collapseItem_fillColor: fillColorCollapse(fieldList, ribbonList),
       collapseItem_stroke: strokeCollapse(),
-      collapseItem_labelstyle: lableCollapse(fieldList),
+      collapseItem_labelstyle: labelCollapse(fieldList),
       collapseItem_other: otherAttributesCollapse(),
     },
   };
