+++ src/components/LayerAttribute/common-schema/label-collapse.ts
