+++ src/components/LayerAttribute/types/index.ts
@@ -1,11 +1,1 @@
-/**
- * 数据列表字段项
- */
-
-export type FieldSelectOptionType = {
-  value: string;
-  lable: string;
-  type?: string;
-  typeName?: string;
-  typeColor?: string;
-};
+export type { FieldSelectOptionType } from '@antv/insight-component';
